# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ILMA API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CurrentUser {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isAdmin: boolean;
}

export interface Bookmark {
  id: number;
  itemType: string;
  itemId: string;
  title: string;
  createdAt: string;
}

export type BookmarkInputItemType = typeof BookmarkInputItemType[keyof typeof BookmarkInputItemType];


export const BookmarkInputItemType = {
  career: 'career',
  domain: 'domain',
  roadmap: 'roadmap',
  exam: 'exam',
  resource: 'resource',
  news: 'news',
  company: 'company',
} as const;

export interface BookmarkInput {
  itemType: BookmarkInputItemType;
  itemId: string;
  title: string;
}

export interface BookmarkToggleResult {
  bookmarked: boolean;
}

export interface ProgressEntry {
  roadmapId: string;
  phase: string;
  topicIndex: number;
  completedAt: string;
}

export type ProgressInputPhase = typeof ProgressInputPhase[keyof typeof ProgressInputPhase];


export const ProgressInputPhase = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface ProgressInput {
  roadmapId: string;
  phase: ProgressInputPhase;
  /** @minimum 0 */
  topicIndex: number;
}

export interface ProgressToggleResult {
  completed: boolean;
}

export interface RecentItem {
  itemType: string;
  itemId: string;
  title: string;
  viewedAt: string;
}

export interface RecentItemInput {
  itemType: string;
  itemId: string;
  title: string;
}

export interface NotificationItem {
  id: number;
  title: string;
  body: string;
  category: string;
  /** @nullable */
  link?: string | null;
  createdAt: string;
  read: boolean;
}

export type NotificationInputCategory = typeof NotificationInputCategory[keyof typeof NotificationInputCategory];


export const NotificationInputCategory = {
  roadmap: 'roadmap',
  job: 'job',
  exam: 'exam',
  news: 'news',
  general: 'general',
} as const;

export interface NotificationInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  body: string;
  category: NotificationInputCategory;
  link?: string;
}

export interface ResourceItem {
  id: number;
  title: string;
  description: string;
  category: string;
  difficulty: string;
  topic: string;
  url: string;
  featured: boolean;
}

export type ResourceInputCategory = typeof ResourceInputCategory[keyof typeof ResourceInputCategory];


export const ResourceInputCategory = {
  book: 'book',
  paper: 'paper',
  dataset: 'dataset',
  software: 'software',
  youtube: 'youtube',
  website: 'website',
  mooc: 'mooc',
  tool: 'tool',
} as const;

export type ResourceInputDifficulty = typeof ResourceInputDifficulty[keyof typeof ResourceInputDifficulty];


export const ResourceInputDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
  all: 'all',
} as const;

export interface ResourceInput {
  /** @minLength 1 */
  title: string;
  description: string;
  category: ResourceInputCategory;
  difficulty?: ResourceInputDifficulty;
  topic: string;
  url: string;
  featured?: boolean;
}

export interface ResourceUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  category?: string;
  difficulty?: string;
  topic?: string;
  url?: string;
  featured?: boolean;
}

export interface NewsArticle {
  id: number;
  title: string;
  summary: string;
  category: string;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  sourceName?: string | null;
  trending: boolean;
  publishedAt: string;
}

export type NewsInputCategory = typeof NewsInputCategory[keyof typeof NewsInputCategory];


export const NewsInputCategory = {
  'ai-healthcare': 'ai-healthcare',
  'medical-devices': 'medical-devices',
  research: 'research',
  fda: 'fda',
  who: 'who',
  healthtech: 'healthtech',
} as const;

export interface NewsInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  summary: string;
  category: NewsInputCategory;
  sourceUrl?: string;
  sourceName?: string;
  trending?: boolean;
  publishedAt?: string;
}

export interface NewsUpdate {
  /** @minLength 1 */
  title?: string;
  summary?: string;
  category?: string;
  sourceUrl?: string;
  sourceName?: string;
  trending?: boolean;
}

export interface NewsletterInput {
  email: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  message: string;
}

export type TrackInputEventType = typeof TrackInputEventType[keyof typeof TrackInputEventType];


export const TrackInputEventType = {
  page_view: 'page_view',
  item_view: 'item_view',
} as const;

export interface TrackInput {
  eventType: TrackInputEventType;
  itemType?: string;
  itemId?: string;
  path?: string;
}

export interface Subscriber {
  id: number;
  email: string;
  createdAt: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  subject: string;
  message: string;
  read: boolean;
  createdAt: string;
}

export interface PopularItem {
  itemType: string;
  itemId: string;
  views: number;
}

export interface AdminStats {
  totalUsers: number;
  totalBookmarks: number;
  subscriberCount: number;
  unreadMessages: number;
  pageViews: number;
  popularItems: PopularItem[];
}

export type ListResourcesParams = {
category?: string;
difficulty?: string;
topic?: string;
search?: string;
};

export type ListNewsParams = {
category?: string;
trending?: boolean;
};