# lib/api-zod/src/generated/types/progressEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ILMA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressEntry {
  roadmapId: string;
  phase: string;
  topicIndex: number;
  completedAt: Date;
}