# lib/api-zod/src/generated/types/recentItemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ILMA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecentItemInput {
  itemType: string;
  itemId: string;
  title: string;
}