# lib/api-zod/src/generated/types/resourceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ILMA API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceInputCategory } from './resourceInputCategory';
import type { ResourceInputDifficulty } from './resourceInputDifficulty';

export interface ResourceInput {
  /** @minLength 1 */
  title: string;
  description: string;
  category: ResourceInputCategory;
  difficulty?: ResourceInputDifficulty;
  topic: string;
  url: string;
  featured?: boolean;
}