# lib/api-zod/src/generated/types/resourceItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ILMA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceItem {
  id: number;
  title: string;
  description: string;
  category: string;
  difficulty: string;
  topic: string;
  url: string;
  featured: boolean;
}