# lib/api-zod/src/generated/types/resourceUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ILMA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  category?: string;
  difficulty?: string;
  topic?: string;
  url?: string;
  featured?: boolean;
}